Checks the status of uwz.at
Gets the map with the actual warnings. As these warnings are shown in different colors (green - no warning, yellow - prewarning, ....) the map is downloaded as image. Then the getPixel is used to get the color for a specific region (use Paint.net before to get the Pixel-Location of the place you life).